backend/requirements.txt
coverage==4.5.1
Django==2.1
django-cors-headers==2.4.0
django-filter==2.0.0
djangorestframework==3.8.2

